Add star_trek test driver for rotation and vertices
The new driver code star_trek.cc creates an arrowhead domain or
'star trek logo' using the intersection of two ellipses. This is
used to test the rotation of boundaruy dofs, in particular at
vertices.

Also hidden in here is a tidy of circular_disc.cc after a messy
merge.