Add otline interface for duplicating and constraining vertices in star_trek.cc
In star_trek.cc, add the outline of code that will be needed to duplicate
nodes in the problem class, also add an incomplete element which will
apply these constraints.

The aim of this is to apply boundary conditions to each node independently
(one for each boudnary) and then constrain the interpolation using the chain
rule.