LARGE: add drivers created to test rotation and corners
This commit message will first outline the new drivers before
explaining major source changes that introduced new features
here.

---------------------------------------------------------------------
New drivers include:
-- rotated_square_upgraded.cc
      copy of rotated square except it upgrades the elements to have
      "curved" boundaries that interpolate the straight edges. This
      was just to help debug curved element issues by comparing
      against unupgraded but rotated elements.

-- sector_corner_modes.cc
      driver to compare the numerical results against analytic corner
      modes of the biharmonic equation on a sector. Uses the .csv
      files found in the Modes_sector directory.

-- square_corner_modes.cc
      driver to compare the numerical results against analytic corner
      modes of the biharmonic equation on a square. Uses the .csv
      files found in the Modes_square directory.
      (similar to sector_corner_modes.cc but without dof rotation or
      element upgrade)

-- triangular_plate.cc
      simple triangular plate driver to test rotation and corner
      constraints on unupgraded elements.

---------------------------------------------------------------------
This commit follows a long period of feature adding, interface
changes and debugging. Recent changes include:
1) Element splitting for elements with two edges on boundaries
2) Node duplication at boundary connections
3) Constraint elements to enforce continuity at duplicated nodes
4) Repair of integration scheme used by p5 elements (no impact on
   drivers)

1)
Each driver with rotated or upgraded boundaries must now use the
function
  void TriangleMeshBase::split_elements_with_multiple_boundary_edges(
    TimeStepper* const& time_stepper_pt)
to ensure no boundary element contains two different edges on
boundaries. This would prevent node duplication at boundary junctions
(see '2)') as well as element upgrade which requires boundary elements
to have exactly one curved edge.

2)
At boundary discontinuities (usually corners, which we assume occur
at junctions between boundaries) we have a discontinuity in the
normal-tangent directions. This normal-tangent coordinate system is
how we apply physical boundary conditions and so on the boundaries
either side of the discontinuity, we require different Hermite dofs.
We achieve this by duplicating the node at the corner so each node can
be in charge of interpolation on each side. This also requires edges
either side of the node to belong to different elements which is
handled by '1)' outlined above. This duplication procedure is not
automated but examples can be found in many drivers such as
'rotated_square.cc' and 'circular_sector.cc'.

3)
By using two different sets of dofs at a corner to interpolate one
variable, we lose the guarentee of continuity that Hermite dofs are
used for. To fix this problem, we use Lagrange multipliers to apply
constraints between the sets of dofs in each node. These constraints
are implemented in a new GeneralisedElement found in
  c1_foeppl_von_karman/foeppl_von_karman_curvable_bell_elements.h
and more details about them can be found in the documentation for
these constraint elements. All the user has to do is -- for each
pair of duplicate nodes -- create a new
  DuplicateNodeConstraintElement(
    Node* const &left_node_pt,
    Node* const &right_node_pt,
    CurvilineGeomObject* const &left_boundary_pt,
    CurvilineGeomObject* const &right_boundary_pt,
    Vector<double> const &left_coord,
    Vector<double> const &right_coord)
whose constructor expects pointers to each of the duplicate nodes
as well as their respective parametrisation and boundary coordinate
(so that normal and tangent data can be retrieved). Also, IMPORTANTLY
AFTER boundary conditions are applied to the FvK boundary elements
BUT BEFORE assign_eqn_numbers() is called, the function
  DuplicateNodeConstraintElement::
    validate_and_pin_redundant_constraints()
must be called in each constraint element. This ensures that the
necessary and sufficient constraints to ensure continuity are applied.
Any more and the system is overconstrained resulting in a singular
jacobian when solving. Any fewer and continuity cannot be ensured.

4)
A long standing bug in the integration scheme used by the p5 upgraded
Bell elements led to many drivers being created and edited for
debugging. The bug was initially believed to be in the elements
themselves however was a result of mixed integration schemes being
used between elements -- incorrect weights led to incorrect
contributions. This should have no lasting effect on these driver
codes but explains why this commit is so large (it follows a hectic
period) and why there may be leftover debugging outputs here.

---------------------------------------------------------------------